tb_history, dto, dao, 예약, 수정, 삭제시 이력내역 추가, 이력내역 검색

+++ src/com/ibm/cof/controller/RsvController/DeleteRsv.java
@@ -1,6 +1,7 @@
 package com.ibm.cof.controller.RsvController;
 
 import java.io.IOException;
+import java.util.ArrayList;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -10,7 +11,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import com.ibm.cof.dao.AdminDAO;
+import com.ibm.cof.dao.HistoryDAO;
 import com.ibm.cof.dao.RsvDAO;
+import com.ibm.cof.dto.HistoryDTO;
+import com.ibm.cof.dto.RsvDTO;
 
 /**
  * Servlet implementation class DeleteRsv
@@ -54,13 +59,40 @@ public class DeleteRsv extends HttpServlet {
 		RsvDAO rdao = new RsvDAO();
 		String password = rdao.getPassword(seq);
 		
+		
+		// 회의실 예약 내역 추가하기 위해 정보 가져오기
+		RsvDTO rdto = rdao.selectBySeq(seq);
+		System.out.println(rdto.getRsv_Date());
+		System.out.println(rdto.getRsv_Start_Time());
+		System.out.println(rdto.getRsv_Title());
+		System.out.println(rdto.getRsv_Site());
+		System.out.println(rdto.getRsv_Confer_Nm());
+		System.out.println(rdto.getRsv_Mem_Nm());
+		System.out.println(rdto.getRsv_Mem_Pn());
+		System.out.println(rdto.getRsv_Mem_Em());
+		System.out.println(rdto.getRsv_Del_Pw());
+		
+		HistoryDAO hdao = new HistoryDAO();
+		HistoryDTO hdto = new HistoryDTO(rdto.getRsv_Date(), rdto.getRsv_Start_Time(), rdto.getRsv_End_Time(), 
+				rdto.getRsv_Title(), rdto.getRsv_Site(), rdto.getRsv_Confer_Nm(), rdto.getRsv_Mem_Nm(),
+				rdto.getRsv_Mem_Pn(), rdto.getRsv_Mem_Em(), rdto.getRsv_Del_Pw(), "삭제");
+	
+		AdminDAO adao = new AdminDAO();
+		
 		// 관리자는 모든 예약 삭제가능
 		if(session.getAttribute("admin") == "yes" && session.getAttribute("project") != "master"){
-			rdao.delete(seq);
-			System.out.println("삭제완료");
+			String project = (String)session.getAttribute("project");
+			String projPassword = adao.selectPassword(project);
+			
+			if(projPassword.equals(del_pw)){
+				rdao.delete(seq);
+				hdao.insert(hdto);
+				System.out.println("삭제완료");
+			}			
 		}
 		else if(password.equals(del_pw)) { //비밀번호 일치여부 확인
 			rdao.delete(seq);
+			hdao.insert(hdto);
 			System.out.println("삭제완료");
 		}
 		else{
